Separate TalkerScreen and TalkerView
